requirements.txt
Flask==2.1
ruamel.yaml>=0.17
boto3>=1.24
cachetools>=5.2
flask-socketio
#eventlet
#gunicorn
flask-login